Inverting summing amplifier with more inputs
An operational amplifier, or op-amp, is a versatile electronic component commonly used in analog circuits for signal processing. It typically has two inputs (inverting and non-inverting) and an output. The inverting input, when connected with certain configurations, allows for the creation of an inverting amplifier. In an inverting configuration, the output signal is an inverted or opposite polarity version of the input signal.